Pass VideoTransformTrack where fps is calculated to WebRTCVideoFrameProducer
